// src/types/types.ts
// Generated by https://quicktype.io

export interface HubsState {
    hubsState: {
        isLoading: boolean;
        hubs: HubsResponse[];
    }
}
export interface HubsResponse {
    uuid:                            string;
    state:                           string;
    category:                        string;
    stage:                           string;
    name:                            string;
    displayName:                     string;
    slug:                            null | string;
    type:                            null | string;
    location:                        null | string;
    externalId:                      null | string;
    recoveredQuantity:               number | null;
    recoveredQuantityUnit:           QuantityUnit | null;
    totalRecoveredQuantity:          number;
    collectionAndSortingParagraph:   null | string;
    pageMode:                        string;
    hubUnassignedRecoveryList:       HubUnassignedRecoveryList[];
    referenceQuantityUnit:           QuantityUnit;
    parentHubName:                   null | string;
    logo:                            CardImage | null;
    cardDescription:                 string;
    cardImage:                       CardImage;
    thankYouNote:                    null | string;
    assignable:                      boolean;
    unassignedQuantityTotal:         number;
    formattedRecoveredQuantity:      null | string;
    formattedTotalRecoveredQuantity: string;
}

export interface CardImage {
    uuid:                string;
    directLink:          string;
    thumbnailDirectLink: string;
    fileName:            string;
    size:                number;
}

export interface HubUnassignedRecoveryList {
    uuid:               string;
    createdAt:          string;
    state:              State;
    unassignedQuantity: number;
    assignedQuantity:   number;
    quantityUnit:       QuantityUnit;
}

export enum QuantityUnit {
    Kg = "KG",
}

export enum State {
    PartiallyAssigned = "PARTIALLY_ASSIGNED",
    Unassigned = "UNASSIGNED",
}